Remove setup for UniqunessValidator. #klass was only used while creating a criteria.
IN PROGRESS

And can be derived from document.class. This change is necessarry for use with rails 4.2+
because of their chain of run_callback methods. At one point it calls super on
our class and sends it #validate_each which in turn calls validate_root. Now I
don't completely grok the changes rails has made, and the path diverges from the
way 4.1 works because our overwrite of #run_callbacks is called during the
validation process.

I want to really understand what the changes affect before use in production.